Eliminate redundant Room TypeConverters
https://developer.android.com/training/data-storage/room/referencing-data#type-converters
