+++ service/src/main/kotlin/com/github/kr328/clash/service/data/Imported.kt
@@ -1,12 +1,10 @@
 package com.github.kr328.clash.service.data
 
 import androidx.room3.Entity
-import androidx.room3.TypeConverters
 import com.github.kr328.clash.service.model.Profile
 import kotlin.uuid.Uuid
 
 @Entity(tableName = "imported", primaryKeys = ["uuid"])
-@TypeConverters(RoomTypeConverters::class)
 data class Imported(
   val uuid: Uuid,
   val name: String,
