+++ service/src/main/kotlin/com/github/kr328/clash/service/data/PendingDao.kt
@@ -4,12 +4,10 @@ import androidx.room3.Dao
 import androidx.room3.Insert
 import androidx.room3.OnConflictStrategy
 import androidx.room3.Query
-import androidx.room3.TypeConverters
 import androidx.room3.Update
 import kotlin.uuid.Uuid
 
 @Dao
-@TypeConverters(RoomTypeConverters::class)
 interface PendingDao {
   @Query("SELECT * FROM pending WHERE uuid = :uuid") suspend fun queryByUUID(uuid: Uuid): Pending?
